seperate fancy.py into two files
