+++ updatelogs/logs082024/08142024.md
@@ -17,4 +17,4 @@ wow, just found a long time bug...
 
 ![alt text](</updatelogs/images/082024/08142024 - 1.png>)
 
-ok, the new `render.py` seems to be working!
+ok, the new `render.py` seems to be working! ok, quick commit before things break... ok, i think the next move is splitting `fancy.py` into two files... (amazing split screen setup! (with a video playing in the background!)) ok, the file should be split now! ok, i guess next is possibly adding in the new label wrapper! again, quick commit!
